% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bat_mix.R
\name{bat_mix}
\alias{bat_mix}
\title{Posterior of mixture of Power or Inverse Batschelet distributions.}
\usage{
bat_mix(th, bat_type = "power", n_comp = 3,
  mu_logprior_fun = function(mu) 0, kp_logprior_fun = function(kp) 0,
  lam_logprior_fun = function(lam) 0, alph_prior_param = rep(1,
  n_comp), fixed_pmat = matrix(NA, n_comp, 4), niter = 1000, ...)
}
\arguments{
\item{th}{Circular observations, either \code{numeric} in radians, or
\code{circular}.}

\item{bat_type}{Either 'inverse' or 'power', the type of distribution to fit.
The two distributions are similar, but the power Batschelet distribution is
computationally much less demanding.}

\item{n_comp}{Integer; Fixed number of components to estimate.}

\item{mu_logprior_fun}{Function; A function with a single argument, which
returns the log of the prior probability of \eqn{\mu}. Defaults to a
uniform prior function.}

\item{kp_logprior_fun}{Function; A function with a single argument, which
returns the log of the prior probability of \eqn{\kappa}. Defaults to a
uniform prior function. In contrast to the other parameters, for
\eqn{\kappa} the constant (uniform) prior is improper.}

\item{lam_logprior_fun}{Function; A function with a single argument, which
returns the log of the prior probability of \eqn{\lambda}. Defaults to a
uniform prior function.}

\item{alph_prior_param}{Integer vector; The mixture weight parameter vector
\eqn{\alpha} is given its conjugate Dirichlet prior. The default is
\code{rep(1, n_comp)}, which is the noninformative uniform prior over the
\code{n_comp} simplex.}

\item{fixed_pmat}{A numeric matrix with \code{n_comp} rows and four columns,
corresponding to \eqn{\mu, \kappa, \lambda, \alpha}, in that order. Any
element that is not \code{NA} in this matrix will be held constant at the
given value and not sampled.}

\item{niter}{Number of iterations to perform MCMC for.}

\item{...}{Further arguments passed to \code{circglmbayes::fitbatmix}.}
}
\value{
Object of type \code{bat_mix_mod}.
}
\description{
Posterior of mixture of Power or Inverse Batschelet distributions.
}
\examples{
bat_mix(rvm(30, 2, 5))

}
