% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bat_posterior.R
\name{bat_posterior}
\alias{bat_posterior}
\title{Posterior of the Power or Inverse Batschelet distribution.}
\usage{
bat_posterior(th, bat_type = "power", mu_logprior_fun = function(mu) 0,
  kp_logprior_fun = function(kp) 0, lam_logprior_fun = function(lam) 0,
  niter = 1000, ...)
}
\arguments{
\item{th}{Circular observations, either \code{numeric} in radians, or
\code{circular}.}

\item{niter}{Number of iterations to perform MCMC for.}

\item{...}{Further arguments passed to \code{circglmbayes::fitbatmix}.}
}
\description{
Posterior of the Power or Inverse Batschelet distribution.
}
\examples{
bat_posterior(rvm(30, 2, 5))

}
