% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{inf_crit}
\alias{inf_crit}
\title{Extract information criteria from fit model object.}
\usage{
inf_crit(x, ...)
}
\arguments{
\item{x}{Fit model object.}

\item{...}{Further arguments.}
}
\value{
A dataframe or list of information criteria.
}
\description{
Extract information criteria from fit model object.
}
