% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{marg_lik}
\alias{marg_lik}
\title{Compute the Marginal likelihood of circular Bayesian models.}
\usage{
marg_lik(x, ...)
}
\arguments{
\item{x}{Fit model object.}

\item{...}{Further arguments.}
}
\value{
Numeric; the log marginal likelihood.
}
\description{
For most models in \code{circbayes}, this uses \code{\link[bridgesampling]{bridge_sampler}}.
}
\examples{
th <- rvm(40, 5, 8)
vm_mod  <- vm_posterior(th)
pn_mod  <- pn_posterior(th)
bat_mod <- bat_posterior(th)
mlvm <- marg_lik(vm_mod)
mlpn <- marg_lik(pn_mod)
mlbt <- marg_lik(bat_mod)

bht_compare(mlvm, mlpn, mlbt)

}
