% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_testing.R
\name{marg_lik_circ_unif}
\alias{marg_lik_circ_unif}
\title{Compute the marginal likelihood of the circular uniform distribution.}
\usage{
marg_lik_circ_unif(th)
}
\arguments{
\item{th}{Vector or data frame with circular observations.}
}
\value{
Numeric.
}
\description{
As this is simply \code{-n * log(2 * pi)}, only the length of \code{th} is
used.
}
\examples{
marg_lik_circ_unif(rvm(20))

}
