% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circ_plot_utilities.R
\name{plot_circbayes_regression}
\alias{plot_circbayes_regression}
\title{Plot a circular model.}
\usage{
plot_circbayes_regression(x, pred_name, fit_params, pred_fun,
  add_data = TRUE, add_fit = TRUE, n_samples = 0,
  alpha_samples = 0.3, add_ci = FALSE, qpts = 100,
  icept_name = "Intercept", ...)
}
\arguments{
\item{...}{}
}
\description{
Plot a circular model.
}
\examples{

}
