% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circ_plot_utilities.R
\name{plot_circbayes_univariate}
\alias{plot_circbayes_univariate}
\title{Plot a circular model.}
\usage{
plot_circbayes_univariate(x, pdf_fun = dvm, polar_coord = TRUE,
  add_data = TRUE, add_fit = TRUE, n_samples = 0, add_ci = FALSE,
  bins = 90, r = 3, ymax = NA, qpts = 100, start = pi/2,
  direction = -1, ...)
}
\arguments{
\item{...}{}
}
\description{
Plot a circular model.
}
\examples{

}
