% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pn_reg.R
\name{pn_reg}
\alias{pn_reg}
\title{Posterior of the von Mises distribution.}
\usage{
pn_reg(formula, data, niter = 1000, thin = 1, burnin = 0, ...)
}
\arguments{
\item{niter}{Number of iterations to perform MCMC for.}

\item{...}{Further arguments passed to \code{circglmbayes::circGLM}.}

\item{th}{Circular observations, either \code{numeric} in radians, or
\code{circular}.}

\item{prior}{Numeric of length 3. Prior parameters for conjugate prior of the
von Mises distribution. The order is \eqn{\mu_0, R_0, c}.}
}
\description{
Posterior of the von Mises distribution.
}
\examples{

}
