% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{posterior_samples}
\alias{posterior_samples}
\title{Generic method to extract posterior (MCMC) samples from a Bayesian model fit
object.}
\usage{
posterior_samples(object, ...)
}
\arguments{
\item{object}{Object to extract samples from.}

\item{...}{}
}
\value{
A matrix of posterior samples.
}
\description{
Generic method to extract posterior (MCMC) samples from a Bayesian model fit
object.
}
\examples{
mod <- vm_posterior(rvm(30, 2, 5), niter = 30)
posterior_samples(mod)
}
