% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{predict_function}
\alias{predict_function}
\title{Obtain the prediction function of a regression object.}
\usage{
predict_function(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Further arguments.}
}
\value{
A function.
}
\description{
Obtain the prediction function of a regression object.
}
