% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular_data_handling.R
\name{print.circrad}
\alias{print.circrad}
\title{Print a \code{circrad} object.}
\usage{
\method{print}{circrad}(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Further arguments passed to \code{print.numeric}}
}
\description{
Print a \code{circrad} object.
}
\examples{
print(circrad(pi/2))

}
