% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bat_posterior.R
\name{rinvbat}
\alias{rinvbat}
\title{Random generation for the Inverse Batschelet distribution.}
\usage{
rinvbat(n, mu = 0, kp = 1, lam = 0)
}
\arguments{
\item{n}{Number of values to sample.}

\item{mu}{Mean direction.}

\item{kp}{Concentration parameter.}

\item{lam}{Lambda, peakedness parameter, between -1 and 1. Positive values
give a peaked density, while negative values give flat-topped densities. .}
}
\value{
Numeric vector of \code{n} samples from the Inverse Batschelet
  distribution, in radians.
}
\description{
Random generation for the Inverse Batschelet distribution.
}
\examples{
hist(rinvbat(1000, .3, 2, .8), breaks = 100)

}
