% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bat_mix.R
\name{rinvbatmix}
\alias{rinvbatmix}
\title{Random generation for the Inverse Batschelet mixture distribution.}
\usage{
rinvbatmix(n, mus = -1:1, kps = c(50, 25, 10), lams = c(-0.4, 0,
  0.6), alphs = c(0.2, 0.2, 0.6))
}
\arguments{
\item{n}{Number of values to sample.}

\item{mus}{Mean directions.}

\item{kps}{Concentration parameters.}

\item{lams}{Lambdas, peakedness parameter, between -1 and 1. Positive values
give a peaked density, while negative values give flat-topped densities. .}

\item{alphs}{Mixture weights.}
}
\value{
Numeric vector of \code{n} samples from the Inverse Batschelet
  distribution, in radians.
}
\description{
Random generation for the Inverse Batschelet mixture distribution.
}
\examples{
hist(rinvbatmix(1000), breaks = 100)

}
