% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_posterior.R
\name{rvm}
\alias{rvm}
\title{Random generation for the von Mises distribution.}
\usage{
rvm(n, mu = 0, kp = 1)
}
\arguments{
\item{n}{Number of values to sample.}

\item{mu}{Mean direction.}

\item{kp}{Concentration parameter.}
}
\value{
Numeric vector of \code{n} samples from the von Mises disttribution,
  in radians.
}
\description{
Random generation for the von Mises distribution.
}
\examples{
rvm(40, 3, 2)

}
