% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_dpm.R
\name{vm_dpm}
\alias{vm_dpm}
\title{Bayesian inference for Dirichlet Process Mixture of Von Mises distributions.}
\usage{
vm_dpm(th, g0Priors = c(0, 0, 1), alphaPriors = c(2, 4),
  niter = 1000, ...)
}
\arguments{
\item{th}{Circular observations, either \code{numeric} in radians, or
\code{circular}.}

\item{niter}{Number of iterations to perform MCMC for.}

\item{...}{Further arguments passed to \code{circglmbayes::fitbatmix}.}
}
\description{
Bayesian inference for Dirichlet Process Mixture of Von Mises distributions.
}
\examples{
vm_dpm(rvm(30, 2, 5))

}
