% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_mix.R
\name{vm_mix}
\alias{vm_mix}
\title{Posterior of mixture of Von Mises distributions.}
\usage{
vm_mix(th, n_comp = 3, mu_logprior_fun = function(mu) 0,
  kp_logprior_fun = function(kp) 0, alph_prior_param = rep(1, n_comp),
  fixed_pmat = matrix(NA, n_comp, 4), niter = 1000, ...)
}
\arguments{
\item{th}{Circular observations, either \code{numeric} in radians, or
\code{circular}.}

\item{n_comp}{Integer; Number of mixture components.}

\item{mu_logprior_fun}{Function; A function with a single argument, which
returns the log of the prior probability of \eqn{\mu}. Defaults to a
uniform prior function.}

\item{kp_logprior_fun}{Function; A function with a single argument, which
returns the log of the prior probability of \eqn{\kappa}. Defaults to a
uniform prior function. In contrast to the other parameters, for
\eqn{\kappa} the constant (uniform) prior is improper.}

\item{alph_prior_param}{Integer vector; The mixture weight parameter vector
\eqn{\alpha} is given its conjugate Dirichlet prior. The default is
\code{rep(1, n_comp)}, which is the noninformative uniform prior over the
\code{n_comp} simplex.}

\item{fixed_pmat}{A numeric matrix with \code{n_comp} rows and four columns,
corresponding to \eqn{\mu, \kappa, \lambda, \alpha}, in that order. Any
element that is not \code{NA} in this matrix will be held constant at the
given value and not sampled.}

\item{niter}{Number of iterations to perform MCMC for.}

\item{...}{Further arguments passed to \code{flexcircmix::fitbatmix}.}
}
\value{
Object of type \code{vm_mix_mod}.
}
\description{
Posterior of mixture of Von Mises distributions.
}
\examples{
vm_mix(rvm(30, 2, 5))

}
