% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_posterior.R
\name{vm_posterior}
\alias{vm_posterior}
\title{Posterior of the von Mises distribution.}
\usage{
vm_posterior(th, prior = c(0, 0, 0), niter = 1000, ...)
}
\arguments{
\item{th}{Circular observations, either \code{numeric} in radians, or
\code{circular}.}

\item{prior}{Numeric of length 3. Prior parameters for conjugate prior of the
von Mises distribution. The order is \eqn{\mu_0, R_0, c}.}

\item{niter}{Number of iterations to perform MCMC for.}

\item{...}{Further arguments passed to \code{circglmbayes::circGLM}.}
}
\description{
Posterior of the von Mises distribution.
}
\examples{
vm_posterior(rvm(30, 2, 5))

}
