% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_reg.R
\name{vm_reg}
\alias{vm_reg}
\title{Posterior of the von Mises distribution.}
\usage{
vm_reg(formula, data, conj_prior = c(0, 0, 0), beta_prior = c(0, 1),
  niter = 1000, r = 2, ...)
}
\arguments{
\item{niter}{Number of iterations to perform MCMC for.}

\item{...}{Further arguments passed to \code{circglmbayes::circGLM}.}

\item{th}{Circular observations, either \code{numeric} in radians, or
\code{circular}.}

\item{prior}{Numeric of length 3. Prior parameters for conjugate prior of the
von Mises distribution. The order is \eqn{\mu_0, R_0, c}.}
}
\description{
Posterior of the von Mises distribution.
}
\examples{
vm_reg(rvm(30, 2, 5))

}
