% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bat_posterior.R
\name{bat_posterior}
\alias{bat_posterior}
\title{Posterior of the Power or Inverse Batschelet distribution.}
\usage{
bat_posterior(th, bat_type = "power", mu_logprior_fun = function(mu) 0,
  kp_logprior_fun = function(kp) 0, lam_logprior_fun = function(lam) 0,
  niter = 1000, ...)
}
\arguments{
\item{th}{Circular observations, either \code{numeric} in radians, or
\code{circular}.}

\item{bat_type}{Either 'inverse' or 'power', the type of distribution to fit.
The two distributions are similar, but the power Batschelet distribution is
computationally much less demanding.}

\item{mu_logprior_fun}{Function; A function with a single argument, which
returns the log of the prior probability of \eqn{\mu}. Defaults to a
uniform prior function.}

\item{kp_logprior_fun}{Function; A function with a single argument, which
returns the log of the prior probability of \eqn{\kappa}. Defaults to a
uniform prior function. In contrast to the other parameters, for
\eqn{\kappa} the constant (uniform) prior is improper.}

\item{lam_logprior_fun}{Function; A function with a single argument, which
returns the log of the prior probability of \eqn{\lambda}. Defaults to a
uniform prior function.}

\item{niter}{Number of iterations to perform MCMC for.}

\item{...}{Further arguments passed to \code{circglmbayes::fitbatmix}.}
}
\value{
Object of type \code{bat_posterior_mod}.
}
\description{
Posterior of the Power or Inverse Batschelet distribution.
}
\examples{
bat_posterior(rvm(30, 2, 5))

}
