% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_testing.R
\name{bht_compare}
\alias{bht_compare}
\title{Bayesian hypothesis tests from marginal likelihoods.}
\usage{
bht_compare(..., add_uniform = FALSE)
}
\arguments{
\item{...}{}
}
\description{
Bayesian hypothesis tests from marginal likelihoods.
}
