% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_testing.R
\name{print.bht_comparison}
\alias{print.bht_comparison}
\alias{bht_compare}
\title{Bayesian hypothesis tests from marginal likelihoods.}
\usage{
\method{print}{bht_comparison}(x, digits = 3, ...)

bht_compare(...)
}
\arguments{
\item{...}{(Named) list of models.}
}
\value{
Object of type \code{'bht_comparison'}.
}
\description{
Bayesian hypothesis tests from marginal likelihoods.
}
\section{Methods (by generic)}{
\itemize{
\item \code{print}: Print method.
}}

\examples{
# Test circular uniformity versus von mises.
th     <- rvm(30, 0, 1)
vm_mod <- vm_posterior(th)

bht_compare(uniform  = marg_lik_circ_unif(th),
            vonmises = marg_lik(vm_mod))

}
