% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_testing.R
\name{print.bht_comparison}
\alias{print.bht_comparison}
\alias{bht_compare}
\title{Print function for Bayesian hypothesis testing.}
\usage{
\method{print}{bht_comparison}(x, digits = 3, ...)

bht_compare(...)
}
\arguments{
\item{x}{Object of type \code{bht_comparison}.}

\item{digits}{Number of digits to print.}

\item{...}{(Named) list of models.}
}
\value{
Object of type \code{'bht_comparison'}.
}
\description{
Print function for Bayesian hypothesis testing.

Bayesian hypothesis tests from marginal likelihoods.
}
\section{Methods (by generic)}{
\itemize{
\item \code{print}: Print method.
}}

\examples{
th <- rvm(100, 3, .2)

vm_mod <- vm_posterior(th, prior = c(0, 0, 1))

cu_ml <- marg_lik_circ_unif(th)
vm_ml <- marg_lik(vm_mod)

bht_compare(uniform = cu_ml, von_mises = vm_ml)
# Test circular uniformity versus von mises.
th     <- rvm(30, 0, 1)
vm_mod <- vm_posterior(th)

bht_compare(uniform  = marg_lik_circ_unif(th),
            vonmises = marg_lik(vm_mod))

}
