% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pn_posterior.R
\name{pn_posterior}
\alias{pn_posterior}
\title{Posterior of the Projected Normal distribution.}
\usage{
pn_posterior(th, niter = 1000, thin = 1, burnin = 0, ...)
}
\arguments{
\item{th}{Circular observations, either \code{numeric} in radians, or
\code{circular}.}

\item{niter}{Number of iterations to perform MCMC for.}

\item{thin}{Integer; Factor of thinning. Setting \code{thin = 1} provides no
thinning.}

\item{burnin}{Integer; Number of iterations to burn.}

\item{...}{Further arguments passed to \code{bpnreg::bpnr}.}
}
\value{
Object of type \code{pn_posterior_mod}.
}
\description{
Posterior of the Projected Normal distribution.
}
\examples{
pn_posterior(rprojnorm(30, 2, 5))

}
