% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pn_reg.R
\name{pn_reg}
\alias{pn_reg}
\title{Bayesian inference for Projected Normal regression.}
\usage{
pn_reg(formula, data, niter = 1000, thin = 1, burnin = 0, ...)
}
\arguments{
\item{niter}{Number of iterations to perform MCMC for.}

\item{...}{Further arguments passed to \code{circglmbayes::circGLM}.}

\item{th}{Circular observations, either \code{numeric} in radians, or
\code{circular}.}

\item{prior}{Numeric of length 3. Prior parameters for conjugate prior of the
von Mises distribution. The order is \eqn{\mu_0, R_0, c}.}
}
\value{
Object of type \code{pn_reg_mod}.
}
\description{
Bayesian inference for Projected Normal regression.
}
\examples{
pn_reg(th ~ ., rvm_reg(20, beta = c(.5, -.2), kp = 50))

}
