% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bat_posterior.R
\name{rinvbat}
\alias{rinvbat}
\alias{dinvbat}
\alias{dpowbat}
\title{Inverse Batschelet distribution}
\usage{
rinvbat(n, mu = 0, kp = 1, lam = 0)

dinvbat(x, mu = 0, kp = 1, lam = 0, log = FALSE)

dpowbat(x, mu = 0, kp = 1, lam = 0, log = FALSE)
}
\arguments{
\item{n}{Number of values to sample.}

\item{mu}{Mean direction.}

\item{kp}{Concentration parameter.}

\item{lam}{Lambda, peakedness parameter, between -1 and 1. Positive values
give a peaked density, while negative values give flat-topped densities.}

\item{x}{Angle in radians for which to evaluate the probability density.}

\item{log}{Logical; whether to return the log probability density.}

\item{x}{Angle in radians for which to evaluate the probability density.}

\item{log}{Logical; whether to return the log probability density.}
}
\value{
Numeric vector of \code{n} samples from the Inverse Batschelet
  distribution, in radians.
}
\description{
Random generation and probability for the Inverse Batschelet Distribution.
}
\section{Functions}{
\itemize{
\item \code{dinvbat}: Density function for Inverse Batschelet.

\item \code{dpowbat}: Density function for Power Batschelet.
}}

\examples{
hist(rinvbat(1000, .3, 2, .8), breaks = 100)

}
