% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pn_posterior.R
\name{rprojnorm}
\alias{rprojnorm}
\title{Random generation for the Projected Normal distribution.}
\usage{
rprojnorm(n, mu1 = 1, mu2 = 1)
}
\arguments{
\item{n}{Number of values to sample.}

\item{mu1}{Mean vector component 1.}

\item{mu2}{Mean vector component 2.}

\item{x}{Angle in radians for which to evaluate the probability density.}
}
\value{
Numeric vector of \code{n} samples from the Projected Normal disttribution,
  in radians.
}
\description{
Random generation for the Projected Normal distribution.
}
\examples{
rprojnorm(40, 3, 2)

}
