% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_posterior.R
\name{rvm}
\alias{rvm}
\alias{dvm}
\title{Von Mises distribution}
\usage{
rvm(n, mu = 0, kp = 1)

dvm(x, mu = 0, kp = 1, log = FALSE)
}
\arguments{
\item{n}{Number of values to sample.}

\item{mu}{Mean direction.}

\item{kp}{Concentration parameter.}

\item{x}{Angle in radians at which to evalu}
}
\value{
Numeric vector of \code{n} samples from the von Mises disttribution,
  in radians.
}
\description{
Random generation and probability density function for the von Mises
distribution. For further von Mises functions, see
\code{\link[circular]{rvonmises}} in package \pkg{circular}.
}
\section{Functions}{
\itemize{
\item \code{dvm}: Probability density function.
}}

\examples{
rvm(40, 3, 2)

}
