% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circ_plot_utilities.R
\name{scale_circular}
\alias{scale_circular}
\alias{scale_x_circular}
\alias{scale_y_circular}
\title{Circular ggplot scales}
\usage{
scale_circular(units = "degrees", nticks = 4, digits = 0,
  limits = c(0, 2 * pi), positive_labels = TRUE,
  scale_function = ggplot2::scale_x_continuous, ...)

scale_x_circular(...)

scale_y_circular(...)
}
\arguments{
\item{units}{The units to display on the axis. See `Details` for the options.}

\item{nticks}{The number of ticks to display. Only relevant for continous
scales.}

\item{digits}{Numeric; The number of digits for continuous scales.}

\item{limits}{Numeric vector; The limits of the plot. Must two numbers that
are  \code{2*pi} apart.}

\item{scale_function}{The scale function from ggplot to use. Either
\code{scale_x_continuous} or \code{scale_y_continuous}. For convenience
this option can be circumvented by directly using \code{scale_x_circular}
or \code{scale_y_circular}.}

\item{...}{Additional arguments to be passed to  \code{scale_x_continuous} or
\code{scale_y_continuous}.}
}
\value{
An object of type \code{ScaleContinuousPosition} that can be added to
  any existing \code{ggplot}.
}
\description{
Extensions of \code{scale_x_continuous} and \code{scale_y_continuous} that
are useful for circular axes.
}
\details{
Behind the screens, radians will be used, but the plot will display
transformed values.

The continuous transformations are \code{"radians"}, \code{"degrees"} and
\code{"hours"}. For these, \code{nticks} and \code{digits} can be set.

The categorical transformations are \code{"cardinal"}, which prints character
labels of the directions (left, right, up, down), and \code{"compass"}, which
prints compass directions (North, South, East, West). Both of these assume
the circular data starts on the right and moves counterclockwise. If not,
appropriate transformations must first be taken, or the labels will be
non-sensical.

Two additional labels are \code{"texpi"} and \code{"texnegpi"}, which print
nice labels for TeX to interpret starting at 0 and \code{-pi} respectively.
This is useful if TikZ is used.
}
\section{Functions}{
\itemize{
\item \code{scale_x_circular}: If circular data is on the x-axis.

\item \code{scale_y_circular}: If circular data is on the y-axis.
}}

\examples{
p <- plot_batmixfit(params = cbind(mu = c(-pi/2, 0, pi/2, pi), kp = 4,
                              lam = c(-.9, .2, .8, 0), alph = .25))

p + scale_x_circular(units = "compass")
p + scale_x_circular(units = "compass", limits = c(-1, 2*pi - 1))

p + scale_x_circular(units = "cardinal")

p + scale_x_circular(units = "texpi")
p + scale_x_circular(units = "texnegpi")

p + scale_x_circular(units = "hours", nticks = 24)

}
