% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_reg.R
\name{vm_reg}
\alias{vm_reg}
\title{Bayesian inference for von Mises regression.}
\usage{
vm_reg(formula, data, prior_b0kp = c(0, 0, 0), prior_btdt = c(0, 1),
  niter = 1000, r = 2, ...)
}
\arguments{
\item{formula}{an object of class "\code{\link{formula}}" (or one that can be
coerced to that class): a symbolic description of the model to be fitted.}

\item{data}{A data frame or matrix containing the circular observations and
predictors.}

\item{prior_b0kp}{Numeric vector of length 3. The prior parameters of the von
Mises conjugate prior for the intercept and concentration parameter.}

\item{prior_btdt}{Numeric vector of length 2. The prior parameters for the
normal prior on the regression coefficients.}

\item{niter}{Number of iterations to perform MCMC for.}

\item{r}{Parameter of the link function, \eqn{g(x) = r * arctan(x)}.}

\item{...}{Further arguments passed to \code{circglmbayes::circGLM}.}
}
\value{
Object of type \code{vm_reg_mod}.
}
\description{
Bayesian inference for von Mises regression.
}
\examples{
dat <- rvm_reg(30)
vm_reg(th ~ ., data = dat)

}
